fix: mixin injection to prevent swap item with offhand while attacking
