+++ src/main/java/yesman/epicfight/mixin/client/MixinClientPacketListener.java
@@ -1,7 +1,5 @@
 package yesman.epicfight.mixin.client;
 
-import net.minecraft.network.protocol.Packet;
-import net.minecraft.network.protocol.game.ServerboundPlayerActionPacket;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -10,7 +8,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.network.protocol.game.ClientboundRespawnPacket;
 import yesman.epicfight.client.events.ClientEvents;
-import yesman.epicfight.client.events.engine.ControlEngine;
 
 @Mixin(value = ClientPacketListener.class)
 public abstract class MixinClientPacketListener {
@@ -18,18 +15,4 @@ public abstract class MixinClientPacketListener {
 	private void epicfight_handleRespawn(ClientboundRespawnPacket clientboundRespawnPacket, CallbackInfo info) {
 		ClientEvents.packet = clientboundRespawnPacket;
 	}
-
-    @Inject(
-            method = "send(Lnet/minecraft/network/protocol/Packet;)V",
-            at = @At("HEAD"),
-            cancellable = true
-    )
-    private void onBeforeSendPacket(Packet<?> packet, CallbackInfo ci) {
-        final boolean isSwapItemWithOffhand = packet instanceof ServerboundPlayerActionPacket actionPacket &&
-                actionPacket.getAction() == ServerboundPlayerActionPacket.Action.SWAP_ITEM_WITH_OFFHAND; 
-        if (isSwapItemWithOffhand && ControlEngine.shouldDisableSwapHandItems()) {
-            // Disables the swap offhand items while in action (e.g., attacking in Epic Fight mode).
-            ci.cancel();
-        }
-    }
 }
