+++ src/main/resources/epicfight.mixins.json
@@ -7,6 +7,7 @@
 		"client.MixinAbstractContainerEventHandler",
 		"client.MixinAgeableListModel",
 		"client.MixinClientPacketListener",
+		"client.MixinClientCommonPacketListenerImpl",
 		"client.MixinCompositeRenderType",
 		"client.MixinEnderDragonRenderer",
 		"client.MixinEntityRenderer",
